+++ ios/AppLovinMAXNativeAdViewManager.m
@@ -18,12 +18,7 @@ RCT_EXPORT_VIEW_PROPERTY(adUnitId, NSString)
 RCT_EXPORT_VIEW_PROPERTY(placement, NSString)
 RCT_EXPORT_VIEW_PROPERTY(customData, NSString)
 RCT_EXPORT_VIEW_PROPERTY(extraParameters, NSDictionary)
-
-// Callback
-RCT_EXPORT_VIEW_PROPERTY(onAdLoadedEvent, RCTDirectEventBlock)
-RCT_EXPORT_VIEW_PROPERTY(onAdLoadFailedEvent, RCTDirectEventBlock)
-RCT_EXPORT_VIEW_PROPERTY(onAdClickedEvent, RCTDirectEventBlock)
-RCT_EXPORT_VIEW_PROPERTY(onAdRevenuePaidEvent, RCTDirectEventBlock)
+RCT_EXPORT_VIEW_PROPERTY(localExtraParameters, NSDictionary)
 
 // Asset views
 RCT_EXPORT_VIEW_PROPERTY(titleView, NSNumber)
@@ -33,6 +28,12 @@ RCT_EXPORT_VIEW_PROPERTY(callToActionView, NSNumber)
 RCT_EXPORT_VIEW_PROPERTY(iconView, NSNumber)
 RCT_EXPORT_VIEW_PROPERTY(optionsView, NSNumber)
 RCT_EXPORT_VIEW_PROPERTY(mediaView, NSNumber)
+
+// Callback
+RCT_EXPORT_VIEW_PROPERTY(onAdLoadedEvent, RCTDirectEventBlock)
+RCT_EXPORT_VIEW_PROPERTY(onAdLoadFailedEvent, RCTDirectEventBlock)
+RCT_EXPORT_VIEW_PROPERTY(onAdClickedEvent, RCTDirectEventBlock)
+RCT_EXPORT_VIEW_PROPERTY(onAdRevenuePaidEvent, RCTDirectEventBlock)
 
 + (BOOL)requiresMainQueueSetup
 {